web/requirements.txt
Django==1.9.2

gunicorn==19.3.0
psycopg2==2.6
redis==2.10.3
django-redis==4.3.0
ipython==4.1.1

# debugging
ipdb

# forms
django-braces==1.8.1

# models
django-model-utils==2.4

# time utils
pytz

# nfl tools
nflgame==1.2.19
